Allow logging from utils
There is a mutual dependency between the `conf` and `utils` modules, currently with the first including the second. Since the root logger is set up in the first,  can't really log anything as the `LOG` object is setup too soon and doesn't write the log calls.

This PR extracts a `log` module that can be included by in  to break the mutual dependency. As a consequence, some functions have to be moved out of `utils` so that dependencies become like:
- `conf` -> `log`
- `conf` -> `utils`
- `log` -> `dateutils` (new module)
- `utils` -> `dateutils`
